Create l2c_2_The Lounge Numbers_list.py
The Gocargo carriers are very much interested in digitizing their customer support wing. Their first initiative is to setup a large live display in the control room which would display the lounge numbers that are assigned with a particular cargo status.
 
The Customers are grouped in waiting lounges based on their cargo statuses. Given a list of the statuses and a cargo status, print the waiting lounge numbers with the given cargo status. If the cargo status is not matching with any of the statuses, then print 0.
 
The following are the different statuses:
Packed
InTransit
Shipped
Out-For-Delivery
Delivered
 
Note: The lounge number starts from 1,2,3 and so on. There is a possiblity that multiple lounges will  accomodate customers from the same cargo status, if there are more number of customers waiting.
 
Input format:
             The first line of the input is a comma separated list of strings, which corresponds to the statuses of the cargos.
             The second line of the input is a string, which corresponds to the given cargo status.
 
Output format:
            Output is a set of integers, which corresponds to the waiting lounge numbers.
 
Sample Input 1:
Packed,InTransit,Packed,Shipped,Out-For-Delivery,Shipped,Delivered
Packed
Sample Output 1:
1
3

Sample Input 2:
Packed,InTransit,Packed,Out-For-Delivery,Delivered
Shipped
Sample Output 2:
0

Explanation for sample 1: In the given sample, the custumers with the status 'Packed', can be accomodated in the waiting lounges 1, 3.