Create l2c_Variables_and_Operators_1_Automatic volume detection.py
The Gocargo carriers are digitizing their logistics department. The first step in their initiative is automatic volume detection of cargos.
 
The device has a high throughput camera that will capture the image of the cargos that are loaded for logistics. From the captured image, the dimensions of the cargo will be extracted and its volume should be estimated. Can you please help Gocargo in writing a piece of code to calculate the volume of the cargos with the dimensions given.
 
Note :
1. Assume the dimensions are positive.
2. The shape of the cargo is cuboid.
 
Input Format :
            Input consists of 3 integers which correspond to length, width and height of the cargo in metres.
 
Output Format :
            Output consists of a single integer, which corresponds to the volume of the cargo in cubic metres.
 
Sample Input 1:
5
4
3
 Sample Output 1:
60

Explanation for Sample :
length(l) = 5
width(w) = 4
height(h) = 3
volume of cuboid(V) = l * w * h = 5 * 4 * 3 = 60