Create l2c_Tuples_1_Commodity Weight.py
The Gocargo carriers are digitizing their Commodity wing. The first step in their intiative is to calculate the total weights of each commodity, as there are multiple pieces in each commodity. Write a program that will read the weights of several commodities and the number of pieces in each commodities and print the total weights of each commodity in the form of tuple. 

For example,
If the weight of a commodity is X and the number of pieces of the commodity is N, then the total weight of the commodity is X*N.

Input Format:
First line of Input consists of integers separated by spaces, which corresponds to the weights of each commodity.
Second line is an integer that corresponds to the number of pieces in each commodity.
Note: All the commodities have the same number of pieces.

Output Format:
Output is a tuple which contains the total weights of each commodity. 




Sample Input :
11 26 30 45 10 8 9 52 33 47 48
12
Sample Output :
(132, 312, 360, 540, 120, 96, 108, 624, 396, 564, 576)

Explanation for the sample:
In each commodity, there are 12 pieces.
Item 1's weight is 11. Therefore, the total weight = 11 * 12 = 132.
Like wise, we have to calculate for all the commodities.