Create l2c_Dictionary Date and Time_2_Tracking the Shipment Dates.py
Radan Logistics is a mid sized Shipping Company known for its Customer oriented delivery services. The Company Management intended to place an information kiosk in their Head Office which would help their Customers to fetch all the desired information with regards to their shipment. One such vital information that Customers would prefer to know is the details of the working days of the Company which would help them track their shipments.

Help the Management write a program for the kiosk, that when given the start day, number of days 'n' and the start date, will output the dates of the 'n' business days (excluding Saturday and Sunday).
 
Input Format :
 The first line of input is a String, the start day.
The second line of the input is the number of days.
Third line is a string, that correponds to the start date of the shipment.
 
Output Format :
Output is list of dates of 'n' working days seperated by new line.
 
 
Sample Input 1 :
Monday
7
03-12-2007
Sample Output 1 :
04-12-2007
05-12-2007
06-12-2007
07-12-2007
10-12-2007
11-12-2007
 
Sample Input 2 :
Thursday
10
25-02-2012
Sample Output 2 :
26-02-2012
29-02-2012
01-03-2012
02-03-2012
03-03-2012
04-03-2012
07-03-2012
08-03-2012
09-03-2012