Create CC_List_1_All Products.py
In Jain Packers and Movers, there are N number of products for shipment from one place to the another place. The Shippers wanted to find the Cartesian product of the cargo id with the price of the cargo. But while printing, they wanted to print the values of the Cartesian product which will not be a multiple of 2.
So help them to implement the above task by writing a program.

Input Format:
The first line consists of a list of integers separated by spaces that corresponds to the list of cargo ids. 
The second line consists list of integers separated by spaces that corresponds to the prices of the cargoes. 

Output Format: 
A line consists of Cartesian products which are not multiple of 2. If there are no numbers to print display "No such Elements in the list" (without quotes). 
Refer Sample Input/Output for formatting specifications.
The bold letters represent the input.

Sample Input/Output 1: 

Enter the Numbers in List-1
1 2 3

Enter the Numbers in List-2

5 10 15

5 15 15 45


Sample Input/Output 2: 

Enter the Numbers in List-1

4 8 2 10

Enter the Numbers in List-2

7 8 9 4

No such Elements in the list


Explanation for Sample 1:
List 1: 1,2,3
List 2: 5,10,15
Cartesian product of the lists is : 5,10,15,10,20,30,15,30,45
The numbers that are not divisible by 2 are 5,15,15,45.