Create CC_List_2_RotateTheList.py
In Jain Packers and Movers, they are going to place N number of cargoes in a container. But the cargoes are not properly balanced with respect to the weight in that container. So one of the Shipper suggested to shift each cargo into its left side, so that all the cargoes will perfectly be balanced in the container.
So help him to shift (rotate) the cargoes to the left by one position within the container, by writing a program.

Input Format:
A line consists of integers, separated by space.

Output Format:
A line consists of integers, after left shift (rotation).
Refer the Sample Input /Output for formatting specifications.
All bold letters are input.

Sample Input/Output: 

Enter the Numbers in List
1 2 3

Before Rotating : 

1 2 3 

After Rotating : 

2 3 1

Explanation for the sample:
input is 1, 2, 3.  After one shift to the left, the order becomes 2 3 1.