Create CC5_String_1_The Chatbot.py
The Gocargo carriers are building a Chatbot to help its Shipping staff in loading section. Most time is spent by the Staff in categorizing the Shipment Commodities based on their nature and load it to the appropriate Cargo as given below.
 
Write a program to design a chatbot that receives text input from the Shipping staff, identifies the type of the Commodities from the text and tells to which Cargo type it is to be loaded.

If the message contains
1. "medicine","tablet","drugs" - belongs to C-Cargo
2. "chocolate","meat","fruit" - belongs to F-Cargo
3. "electronics","mobile","PC" - belongs to E-Cargo

Note: Assume that the Cargoes will belong only to one of the above mentioned 3 types.

Input format :
Input is an array of strings, the text input given to the Shipping Staff.
 
Output format :
Output is a string, the Cargo type where in which the commodities to be loaded.
 

Sample Input 1: 
The package contains some crispy waffer covered with original Belgium chocolate.
Sample Output 1:
F-Cargo

Explanation for sample:
The input string contains the word 'chocolate' and hence belongs to F-Cargo.