Create CC6_Tuples_1_Duplicate Cargo.py
The Gocargo carriers are into automating every part of their logistics wing. The Container loading machine mis-operates that its display shows duplicate cargo numbers in the container.  An anti-duplicate program is to be programmed on the same machine such that it removes the duplicates and prints the actual cargo numbers that are inside the container.
 
Given the cargo numbers, remove the duplicate cargo in the container and display them.
 
Input Format :
Input consists of integers, a list of cargo numbers.
 
Output Format :
Output is a tuple, that has the list of cargo numbers without duplicates.

 Note: Print the output in the same order as present in the input.

Sample Input:
1,2,3,5,5,6,1


Sample Output:
(1, 2, 3, 5, 6)

Explanation for sample:
The input item numbers is 1,2,3,5,5,6,1.
The final list after removing duplication is  (1, 2, 3, 5, 6).