Create L2C_Functions_1_Loading-Expressions.py
The Gocargo are automating their testing works of the logistics wing too. They have planned to test if the container loading works correctly or not.
Given the process involved in loading, in the form of expression, evaluate if the container loader has worked correct or not.
 
Conventions
'{' for cargos, '[' for containers and '(' for boxes.

Rules:
* cargo cannot be inside another cargo or container or box.
* container cannot be inside another container or box.
* box can be inside another box.
 
Please check whether the container loader has worked correctly.
 
Sample Input 1 :

{[2+3+4+(2+1)+4+(2+4)]+[3+4]}+{[2+4+5]}

 Sample Output 1:
Valid
 
Sample Input 2 :
{[2+3+4+(2+1)+4+(2+4)]+3+4]}+{[2+4+5]}}
Sample Output 2:
Invalid

Explanation for sample 1:
There are two cargos.
In the first cargo, there are two containers.
          The first container contains two boxes, plus some other items.
          The second container contains no boxes, but two items.
In the second cargo, there is one container with 3 items.
It follows the rules. Therefore, the container loader has worked correctly.

Explanation for sample 2:
The unbalanced brackets shows that the container loader has not worked correctly.