Create L2C_Functions_2_Palindromic_Cargo.py
The Gocargo carriers are celebrating their digitization process, which has given them a lot of benefits and ease of work for their employees. In account of this initiative they are offering their customers to ship their cargoes free of cost, if and only if the cargo names are palindromic.
 
Given a cargo name, check whether the cargo name is a palindrome or not. Consider only alphabets while checking, remove special characters and spaces.
 
Note : Define a function named check Palindrome() with a string argument.
 
Input Format :
Input consists of a string, the cargo's name. Input is case - insensitive.

Output Format:
Output is a string which contains "Yes" or "No".
 
Sample Input 1:
!!Race !Car!!
Sample Output 1:
Yes
 
Sample Input 2:
Racer Car
Sample Output 2:
No

Explanation For Sample 1:
Cargo name: !!Race !Car!!
After removing special characters and spaces, the name is "RaceCar".
Now, making the string case-insensitive, it becomes "racecar", which is a palindrome.