Create CC9_Function_1_Overlapping Cities.py
The GoCarriers Shipping Company decided to expand their delivery services to two new ports. The Shipping Agents of the respective ports outlined the list of cities that the Company's Cargoes has to go across to reach the two new ports for shipping goods.

The Shipping Company has to now ensure that the same cities are not listed under both the ports' traversal path, i.e., the cities in both the list shouldn't overlap. They wanted your help to write a block of code that will identify if the two given list of cities that the Cargoes travel through, overlap or not.

Input Format :
The first line of the input consists of the list of cities separated by commas, given by the Agent to reach the first port.
Second line of the input consists of the list of cities separated by commas, given by the Agent to reach the second port.

 
Output Format :
The first two lines of output are the list of cities of the two ports seperated by new line.
The third line is a string, which is either "Overlapping" or "Not Overlapping"
 

Note: The data are not case-sensitive.

Sample Input 1 :
India,Australia,Switzerland,America
Australia,SriLanka
 
Sample Output 1 :
['India', 'Australia', 'Switzerland', 'America']
['Australia', 'SriLanka']
Overlapping
 
 
Sample Input 2 :
coimbatore,chennai,delhi
Mysore,Bangalore,Hubli,Sirsi
 
Sample Output 2 :
['coimbatore', 'chennai', 'delhi']
['Mysore','Bangalore','Hubli','Sirsi']
Non Overlapping