Create L2C_Files_2_Heaviest-Cargos.py
Cargos details like weight, height, dimensions are often vital when loading them into the Containers, as uneven weights in the Containers would result in severe damages of the goods. The Shipment System should provide a functionality to the Users based on the Cargo weights.

Write a program that will get a list of cargos with their weights from the User and display 'n' Cargos that weighs the most.
 
Read inputs from the file and display the output in the console.
 
Note : read from the file readlines.txt


Input File Format :
List of cargos along with their weights is stored in a file in the increasing order of weights. Find below the file format.
 
Aerosol - 38
Churn - 50
Plank - 55
Slipsheet - 75
Jerrican - 95
Girder - 100
 
Console Input format :
First line of the input at console is an integer that corresponds to the value of "n".
 
Output Format :
Output is a list of "n" items which is read from the input file, in the increasing order of weights.
 

Sample Input and Output:
Enter the n value :
3
Slipsheet - 75 
Jerrican - 95 
Girder - 100