Create L2C_Files_3_XML-to-CSV.py
Naveen is owning an electronics shop. He has an XML file that contains all the products that is present in his shop. He wants to tabulate his items in the excel. He found out that a CSV file can be easily represented as an excel. Write a program to help naveen by reading an XML file and printing it in CSV format with heading.

click here to download the XML file


Input Format:

Input is a XML file. (file name should be product.xml).  The file is already been created in the platform.

Example: 

<productList>

<product>

<id>1</id>

<productName>Mobile</productName>

<cost>6000</cost>

<weight>200</weight>

</product>

<product>

<id>2</id>

<productName>Laptop</productName>

<cost>60000</cost>

<weight>2000</weight>

</product>

</productList>

 


Output Format:

Print the data in CSV format which is in XML file.(file name should be product.csv)

Example: 

id,productName,cost,weight

1,Mobile,6000,200

2,Laptop,60000,2000